bear: Move application code to separate compilation unit
